Struct symbol_table::indexing::Ref

pub struct Ref<T> {
    // some fields omitted
}

Wrapper for a raw pointer which lets us treat it like a reference.

You are strongly discouraged from exposing this type directly in your data structures. This type is essentially a giant footgun. In particular:

No safety checks or lifetimes protect this reference, so a Ref<T> may be invalidated without warning. (You may use a Ref<T> safely by ensuring that the references passed to Ref<T>::new() will never be dropped before the wrappers. A good example of when you'd be able to do this is in in a struct that has Ref<T> references into a data structure that it also owns.)

The impls for Debug, Eq, Hash, Ord, PartialEq, and PartialOrd all dereference the raw pointer that this structure wraps. As a result, a Ref<T> must be removed from any data structures that make use of any of those interfaces before it is invalidated.

Ref<T> wraps a value of type *const T, which is not usually Send or Sync. This restriction is overridden for a Ref<T> wrapper so that data structures which encapsulate it may themselves be Send or Sync. This makes it the responsibility of data structures using such wrappers to satisfy the contracts of those types.