Struct symbol_table::Symbol

pub struct Symbol<T, D> where D: SymbolId {
    // some fields omitted
}

A table entry that associates an instance of T with an atomic symbol.

Types T should not be mutated by any means once they are associated with a SymbolId and stored in a Table. Doing so may invalidate any caching or indexing that is done on top of the table.

Methods

impl<T, D> Symbol<T, D> where D: SymbolId

fn id(&self) -> &D

Returns the symbol's ID.

fn data(&self) -> &T

Returns a reference to the symbol's data.

A Symbol<T> that is owned by a Table does not move in memory as long as it is not dropped from the table. As a result, you may retain a raw pointer to this data and dereference it as long as its parent Symbol<T> is not dropped.