Enum symbol_table::indexing::Insertion

pub enum Insertion<T> {
    Present(T),
    New(T),
}

Indicates whether the result of a symbol lookup had to create a new table entry.

Variants
Present(T)

Result came from item that was already present in table.

New(T)

Result came from item that was not present in table, and a new entry was created as a side effect.

Methods

impl<T> Insertion<T>

fn map<F, X>(&self, f: F) -> Insertion<X> where F: FnOnce(&T) -> X